+++ js/src/setup-mc/setup-stepper/setup-free-listings/shipping-rate/shipping-rate-setup/countries-price-input/edit-rate-button/edit-rate-modal/index.js
@@ -13,7 +13,7 @@ import { STORE_KEY } from '../../../../../../../../data';
 import AppModal from '../../../../../../../../components/app-modal';
 import AppInputControl from '../../../../../../../../components/app-input-control';
 import VerticalGapLayout from '../../../../../components/vertical-gap-layout';
-import AudienceCountrySelect from '.~/components/audience-country-select';
+import AudienceCountrySelect from '../../../../../components/audience-country-select';
 import './index.scss';
 
 const EditRateModal = ( props ) => {
